+++ src/schemathesis/runner/impl/solo.py
@@ -15,7 +15,7 @@ class SingleThreadRunner(BaseRunner):
 
     request_tls_verify: Union[bool, str] = attr.ib(default=True)  # pragma: no mutate
 
-    def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+    def _run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         auth = get_requests_auth(self.auth, self.auth_type)
         with get_session(auth) as session:
             yield from self._run_tests(
@@ -37,7 +37,7 @@ class SingleThreadRunner(BaseRunner):
 
 @attr.s(slots=True)  # pragma: no mutate
 class SingleThreadWSGIRunner(SingleThreadRunner):
-    def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+    def _run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         yield from self._run_tests(
             wsgi_test,
             self.hypothesis_settings,
@@ -56,7 +56,7 @@ class SingleThreadWSGIRunner(SingleThreadRunner):
 
 @attr.s(slots=True)  # pragma: no mutate
 class SingleThreadASGIRunner(SingleThreadRunner):
-    def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+    def _run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         yield from self._run_tests(
             asgi_test,
             self.hypothesis_settings,