+++ src/schemathesis/runner/impl/threadpool.py
@@ -168,7 +168,7 @@ class ThreadPoolRunner(BaseRunner):
     workers_num: int = attr.ib(default=2)  # pragma: no mutate
     request_tls_verify: Union[bool, str] = attr.ib(default=True)  # pragma: no mutate
 
-    def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+    def _run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         """All events come from a queue where different workers push their events."""
         tasks_queue = self._get_tasks_queue()
         # Events are pushed by workers via a separate queue